Revert "PopoverWithMeasuredContent optimization"
